Feature red slip (#36)
計上済み伝票の訂正、キャンセルを行う

+++ src/main/java/com/showka/service/crud/u05/UriageCrudServiceImpl.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import com.showka.domain.KokyakuDomain;
 import com.showka.domain.UriageDomain;
 import com.showka.domain.UriageMeisaiDomain;
+import com.showka.domain.UriageRirekiListDomain;
 import com.showka.domain.builder.UriageDomainBuilder;
 import com.showka.entity.MKokyaku;
 import com.showka.entity.TUriage;
@@ -20,8 +21,10 @@ import com.showka.repository.i.MKokyakuRepository;
 import com.showka.repository.i.TUriageMeisaiRepository;
 import com.showka.repository.i.TUriageRepository;
 import com.showka.service.crud.u01.i.KokyakuCrudService;
+import com.showka.service.crud.u05.i.UriageCancelCrudService;
 import com.showka.service.crud.u05.i.UriageCrudService;
 import com.showka.service.crud.u05.i.UriageMeisaiCrudService;
+import com.showka.service.crud.u05.i.UriageRirekiCrudService;
 import com.showka.value.TaxRate;
 import com.showka.value.TheDate;
 
@@ -43,38 +46,26 @@ public class UriageCrudServiceImpl implements UriageCrudService {
 	@Autowired
 	private KokyakuCrudService kokyakuCrudService;
 
+	@Autowired
+	private UriageRirekiCrudService uriageRirekiCrudService;
+
+	@Autowired
+	private UriageCancelCrudService uriageCancelCrudService;
+
 	@Override
 	public void save(UriageDomain domain) {
 
-		// pk
-		TUriagePK pk = new TUriagePK();
-		pk.setDenpyoNumber(domain.getDenpyoNumber());
-		pk.setKokyakuId(domain.getKokyaku().getRecordId());
-
-		// 売上Entity
-		TUriage e = repo.findById(pk).orElse(new TUriage());
-		e.setHanbaiKubun(domain.getHanbaiKubun().getCode());
-		e.setPk(pk);
-		e.setShohizeiritsu(domain.getShohizeiritsu().getRate().doubleValue());
-		e.setUriageDate(domain.getUriageDate().toDate());
-		e.setRecordId(domain.getRecordId());
-		e.setVersion(domain.getVersion());
+		// entity
+		TUriage e = getEntityFromDomain(domain);
 
 		// save
 		repo.saveAndFlush(e);
 
-		// delete old
-		List<UriageMeisaiDomain> oldMeisaiList = uriageMeisaiCrudService.getDomainList(domain.getRecordId());
-		for (UriageMeisaiDomain old : oldMeisaiList) {
-			if (!domain.getUriageMeisai().contains(old)) {
-				uriageMeisaiCrudService.delete(old);
-			}
-		}
+		// save 履歴
+		uriageRirekiCrudService.save(domain);
 
-		// save 明細
-		for (UriageMeisaiDomain meisai : domain.getUriageMeisai()) {
-			uriageMeisaiCrudService.save(meisai);
-		}
+		// 明細更新
+		uriageMeisaiCrudService.overrideList(domain.getUriageMeisai());
 
 		// set 顧客
 		MKokyaku kokyaku = kokyakuRepo.findByRecordId(domain.getKokyaku().getRecordId());
@@ -119,6 +110,7 @@ public class UriageCrudServiceImpl implements UriageCrudService {
 		b.withRecordId(e.getRecordId());
 		b.withShohizeiritsu(new TaxRate(e.getShohizeiritsu()));
 		b.withUriageDate(new TheDate(e.getUriageDate()));
+		b.withKeijoDate(new TheDate(e.getKeijoDate()));
 		b.withUriageMeisai(uriageMeisai);
 		b.withVersion(e.getVersion());
 
@@ -132,15 +124,73 @@ public class UriageCrudServiceImpl implements UriageCrudService {
 
 	@Override
 	public void delete(UriageDomain domain) {
+		// 売上履歴削除
+		uriageRirekiCrudService.delete(domain);
 
-		domain.getUriageMeisai().forEach(meisai -> {
-			uriageMeisaiCrudService.delete(meisai);
-		});
+		// 残った履歴取得
+		String uriageId = domain.getRecordId();
+		UriageRirekiListDomain rirekiList = uriageRirekiCrudService.getUriageRirekiList(uriageId);
 
+		// 計上済みの履歴があれば、それで売上を上書きし直す。
+		// ない場合は、売上データ自体をすべて削除して終了。
+		if (rirekiList.getList().size() > 0) {
+			UriageDomain domainForOverride = rirekiList.getNewest();
+			UriageDomainBuilder b = new UriageDomainBuilder();
+			// ここで排他制御
+			b.withVersion(domain.getVersion());
+			this.save(b.apply(domainForOverride));
+		} else {
+			// 明細削除
+			domain.getUriageMeisai().forEach(meisai -> {
+				uriageMeisaiCrudService.delete(meisai);
+			});
+			// 売上削除
+			TUriagePK pk = new TUriagePK();
+			pk.setDenpyoNumber(domain.getDenpyoNumber());
+			pk.setKokyakuId(domain.getKokyaku().getRecordId());
+			this.delete(pk, domain.getVersion());
+		}
+	}
+
+	@Override
+	public void cancel(UriageDomain domain) {
+		// delete meisai if exists
+		uriageMeisaiCrudService.deleteAll(domain.getRecordId());
+
+		// save entity
+		TUriage e = getEntityFromDomain(domain);
+		repo.saveAndFlush(e);
+
+		// cancel 履歴
+		uriageRirekiCrudService.cancel(domain);
+
+		// cancel
+		uriageCancelCrudService.save(domain);
+	}
+
+	/**
+	 * Domain -> Entity
+	 * 
+	 * @param domain
+	 *            売上ドメイン
+	 * @return 売上エンティティ
+	 */
+	private TUriage getEntityFromDomain(UriageDomain domain) {
+		// pk
 		TUriagePK pk = new TUriagePK();
 		pk.setDenpyoNumber(domain.getDenpyoNumber());
 		pk.setKokyakuId(domain.getKokyaku().getRecordId());
-		this.delete(pk, domain.getVersion());
+
+		// 売上Entity
+		TUriage e = repo.findById(pk).orElse(new TUriage());
+		e.setHanbaiKubun(domain.getHanbaiKubun().getCode());
+		e.setPk(pk);
+		e.setShohizeiritsu(domain.getShohizeiritsu().getRate().doubleValue());
+		e.setUriageDate(domain.getUriageDate().toDate());
+		e.setKeijoDate(domain.getKeijoDate().toDate());
+		e.setRecordId(domain.getRecordId());
+		e.setVersion(domain.getVersion());
+		return e;
 	}
 
 }
